+++ data/admin/historyauthenticationmethod.py
@@ -1,0 +1,2 @@
+class AuthenticationMethodHistoricalRecordsMixin:
+    history_list_display = ["authentication_method"]
