+++ data/models/__init__.py
@@ -1,4 +1,5 @@
 from .user import User  # noqa: F401
+from .historyauthenticationmethod import AuthenticationMethodHistoricalRecords  # noqa: F401
 from .canteen import Canteen, CanteenImage  # noqa: F401
 from .diagnostic import Diagnostic  # noqa: F401
 from .sector import Sector  # noqa: F401
